repair until queue is empty
